Use SQL_ATTR_DBC_DEFAULT_LIB for cur schema
The SQL_ATTR_DBC_DEFAULT_LIB IBM i-specific CLI attribute does the same
thing as SET CURRENT SCHEMA, so we can define SQL_ATTR_CURRENT_SCHEMA
as it and reduce the code path differences between LUW and IBM i. In
addition, we can use this from connect_helper, but we need to make sure
we do not pass an empty string/None or an invalid attribute error will
be sent to the job log.